backend-agent/requirements.txt
# FastAPI e servidor (Python 3.13 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-multipart>=0.0.12

# LlamaIndex para SQL e RAG (versões compatíveis com Python 3.13)
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0

# LangChain para conversação (versões mais recentes)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Banco de dados
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36
asyncpg>=0.30.0

# OpenAI
openai>=1.54.0

# Utilidades
httpx>=0.27.0
aiohttp>=3.11.0
